projects: Use global taskPool to avoid thread exhaustion
We're seeing our summit instance segfault in deployment due to thread
exhaustion. In short, every update new threads are made but not actually
disposed, due to a shortcoming in the design of `executeInNewThread` whereby
neither `pthread_join` or `pthread_detach` are called until the TLS module
destructor.

To mitigate this, use the fixed global pool and avoid overcommiting the
resources and prevent further thread leaks.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>